requirements.txt
matplotlib==3.0.2
numpy==1.15.4
torch==1.0.0
onnx==1.3.0
git+https://github.com/pytorch/pytorch.git#egg=torch
git+https://github.com/onnx/onnx.git#egg=onnx
torchvision==0.2.1
Pillow==5.3.0
